fix: bug cache not invalidating
The cache should invalidate when we update the emsch_env variable!

+++ src/Helper/Cache/CacheHelper.php
@@ -38,6 +38,10 @@ final class CacheHelper
             return null;
         }
 
+        if ($cachedContentType->getEnvironment() !== $contentType->getEnvironment()) {
+            return null; //update on environment
+        }
+
         if ($cachedContentType->getCacheValidityTag() !== $contentType->getCacheValidityTag()) {
             $this->cache->deleteItem($contentType->getCacheKey());
